removed enabled check, added getAccountStats api, split JetStream / JetStreamManagement so js cannot be cast to jsm
