+++ src/main/java/io/nats/client/impl/NatsJetStream.java
@@ -1,223 +1,26 @@
 package io.nats.client.impl;
 
 import io.nats.client.*;
-import io.nats.client.api.*;
-import io.nats.client.support.NatsJetStreamConstants;
+import io.nats.client.api.AckPolicy;
+import io.nats.client.api.ConsumerConfiguration;
+import io.nats.client.api.ConsumerInfo;
+import io.nats.client.api.PublishAck;
 
 import java.io.IOException;
 import java.time.Duration;
-import java.util.List;
 import java.util.concurrent.CompletableFuture;
 
-import static io.nats.client.support.ApiConstants.SEQ;
-import static io.nats.client.support.JsonUtils.simpleMessageBody;
 import static io.nats.client.support.Validator.*;
 
-public class NatsJetStream implements JetStream, JetStreamManagement, NatsJetStreamConstants {
+public class NatsJetStream extends NatsJetStreamImplBase implements JetStream {
 
-    private final NatsConnection conn;
-    private final JetStreamOptions jso;
-
-    // ----------------------------------------------------------------------------------------------------
-    // Create / Init
-    // ----------------------------------------------------------------------------------------------------
-    NatsJetStream(NatsConnection connection, JetStreamOptions jsOptions) throws IOException {
-        conn = connection;
-        jso = JetStreamOptions.builder(jsOptions).build(); // builder handles null
-
-        checkEnabled();
-    }
-
-    private void checkEnabled() throws IOException {
-        try {
-            Message respMessage = makeRequest(JSAPI_ACCOUNT_INFO, null, jso.getRequestTimeout());
-            if (respMessage == null) {
-                throw new IllegalStateException("JetStream is not enabled.");
-            }
-
-            AccountStatistics stats = new AccountStatistics(respMessage);
-            if (stats.getErrorCode() == 503) {
-                throw new IllegalStateException(stats.getDescription());
-            }
-        } catch (InterruptedException e) {
-            throw new IOException(e);
-        }
-    }
-
-    // ----------------------------------------------------------------------------------------------------
-    // Manage
-    // ----------------------------------------------------------------------------------------------------
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public StreamInfo addStream(StreamConfiguration config) throws IOException, JetStreamApiException {
-        return addOrUpdateStream(config, JSAPI_STREAM_CREATE);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public StreamInfo updateStream(StreamConfiguration config) throws IOException, JetStreamApiException {
-        return addOrUpdateStream(config, JSAPI_STREAM_UPDATE);
-    }
-
-    private StreamInfo addOrUpdateStream(StreamConfiguration config, String template) throws IOException, JetStreamApiException {
-        if (config == null) {
-            throw new IllegalArgumentException("Configuration cannot be null.");
-        }
-        String streamName = config.getName();
-        if (nullOrEmpty(streamName)) {
-            throw new IllegalArgumentException("Configuration must have a valid stream name");
-        }
-
-        String subj = String.format(template, streamName);
-        Message resp = makeRequestResponseRequired(subj, config.toJson().getBytes(), jso.getRequestTimeout());
-        return new StreamInfo(resp).throwOnHasError();
-    }
-
-    @Override
-    public boolean deleteStream(String streamName) throws IOException, JetStreamApiException {
-        String subj = String.format(JSAPI_STREAM_DELETE, streamName);
-        Message resp = makeRequestResponseRequired(subj, null, jso.getRequestTimeout());
-        return new SuccessApiResponse(resp).throwOnHasError().getSuccess();
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public StreamInfo getStreamInfo(String streamName) throws IOException, JetStreamApiException {
-        String subj = String.format(JSAPI_STREAM_INFO, streamName);
-        Message resp = makeRequestResponseRequired(subj, null, jso.getRequestTimeout());
-        return new StreamInfo(resp).throwOnHasError();
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public PurgeResponse purgeStream(String streamName) throws IOException, JetStreamApiException {
-        String subj = String.format(JSAPI_STREAM_PURGE, streamName);
-        Message resp = makeRequestResponseRequired(subj, null, jso.getRequestTimeout());
-        return new PurgeResponse(resp).throwOnHasError();
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public ConsumerInfo addOrUpdateConsumer(String streamName, ConsumerConfiguration config) throws IOException, JetStreamApiException {
-        validateStreamName(streamName, true);
-        validateNotNull(config, "Config");
-        validateNotNull(config.getDurable(), "Durable");
-        return addOrUpdateConsumerInternal(streamName, config);
-    }
-
-    private ConsumerInfo addOrUpdateConsumerInternal(String streamName, ConsumerConfiguration config) throws IOException, JetStreamApiException {
-        String durable = config.getDurable();
-        String requestJSON = new ConsumerCreateRequest(streamName, config).toJson();
-
-        String subj;
-        if (durable == null) {
-            subj = String.format(JSAPI_CONSUMER_CREATE, streamName);
-        } else {
-            subj = String.format(JSAPI_DURABLE_CREATE, streamName, durable);
-        }
-        Message resp = makeRequestResponseRequired(subj, requestJSON.getBytes(), conn.getOptions().getConnectionTimeout());
-        return new ConsumerInfo(resp).throwOnHasError();
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public boolean deleteConsumer(String streamName, String consumer) throws IOException, JetStreamApiException {
-        String subj = String.format(JSAPI_CONSUMER_DELETE, streamName, consumer);
-        Message resp = makeRequestResponseRequired(subj, null, jso.getRequestTimeout());
-        return new SuccessApiResponse(resp).throwOnHasError().getSuccess();
-    }
-
-    @Override
-    public ConsumerInfo getConsumerInfo(String streamName, String consumer) throws IOException, JetStreamApiException {
-        String subj = String.format(JSAPI_CONSUMER_INFO, streamName, consumer);
-        Message resp = makeRequestResponseRequired(subj, null, jso.getRequestTimeout());
-        return new ConsumerInfo(resp).throwOnHasError();
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public List<String> getConsumerNames(String streamName) throws IOException, JetStreamApiException {
-        return getConsumerNames(streamName, null);
-    }
-
-    // TODO FUTURE resurface this api publicly when server supports
-    // @Override
-    private List<String> getConsumerNames(String streamName, String filter) throws IOException, JetStreamApiException {
-        String subj = String.format(JSAPI_CONSUMER_NAMES, streamName);
-        ConsumerNamesReader cnr = new ConsumerNamesReader();
-        while (cnr.hasMore()) {
-            Message resp = makeRequestResponseRequired(subj, cnr.nextJson(filter), jso.getRequestTimeout());
-            cnr.process(resp);
-        }
-        return cnr.getStrings();
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public List<ConsumerInfo> getConsumers(String streamName) throws IOException, JetStreamApiException {
-        String subj = String.format(JSAPI_CONSUMER_LIST, streamName);
-        ConsumerListReader clg = new ConsumerListReader();
-        while (clg.hasMore()) {
-            Message resp = makeRequestResponseRequired(subj, clg.nextJson(), jso.getRequestTimeout());
-            clg.process(resp);
-        }
-        return clg.getConsumers();
-    }
-
-    @Override
-    public List<String> getStreamNames() throws IOException, JetStreamApiException {
-        StreamNamesReader snr = new StreamNamesReader();
-        while (snr.hasMore()) {
-            Message resp = makeRequestResponseRequired(JSAPI_STREAMS, snr.nextJson(), jso.getRequestTimeout());
-            snr.process(resp);
-        }
-        return snr.getStrings();
-    }
-
-    @Override
-    public List<StreamInfo> getStreams() throws IOException, JetStreamApiException {
-        StreamListReader slg = new StreamListReader();
-        while (slg.hasMore()) {
-            Message resp = makeRequestResponseRequired(JSAPI_STREAM_LIST, slg.nextJson(), jso.getRequestTimeout());
-            slg.process(resp);
-        }
-        return slg.getStreams();
-    }
-
-    @Override
-    public MessageInfo getMessage(String streamName, long seq) throws IOException, JetStreamApiException {
-        String subj = String.format(JSAPI_MSG_GET, streamName);
-        Message resp = makeRequestResponseRequired(subj, simpleMessageBody(SEQ, seq), jso.getRequestTimeout());
-        return new MessageInfo(resp).throwOnHasError();
-    }
-
-    @Override
-    public boolean deleteMessage(String streamName, long seq) throws IOException, JetStreamApiException {
-        String subj = String.format(JSAPI_MSG_DELETE, streamName);
-        Message resp = makeRequestResponseRequired(subj, simpleMessageBody(SEQ, seq), jso.getRequestTimeout());
-        return new SuccessApiResponse(resp).throwOnHasError().getSuccess();
+    public NatsJetStream(NatsConnection connection, JetStreamOptions jsOptions) throws IOException {
+        super(connection, jsOptions);
     }
 
     // ----------------------------------------------------------------------------------------------------
     // Publish
-    // ---------------------------------------------------------------------------------------------NatsJsPullSub-------
+    // ----------------------------------------------------------------------------------------------------
 
     /**
      * {@inheritDoc}
@@ -597,7 +400,7 @@ public class NatsJetStream implements JetStream, JetStreamManagement, NatsJetStr
         }
     }
 
-    private String lookupStreamBySubject(String subject) throws IOException, JetStreamApiException {
+    protected String lookupStreamBySubject(String subject) throws IOException, JetStreamApiException {
         String streamRequest = String.format("{\"subject\":\"%s\"}", subject);
         StreamNamesReader snr = new StreamNamesReader();
         Message resp = makeRequestResponseRequired(JSAPI_STREAMS, streamRequest.getBytes(), jso.getRequestTimeout());
@@ -608,7 +411,7 @@ public class NatsJetStream implements JetStream, JetStreamManagement, NatsJetStr
         return snr.getStrings().get(0);
     }
 
-    private static class AutoAckMessageHandler implements MessageHandler {
+    protected static class AutoAckMessageHandler implements MessageHandler {
         MessageHandler userMH;
 
         // caller must ensure userMH is not null
@@ -625,43 +428,5 @@ public class NatsJetStream implements JetStream, JetStreamManagement, NatsJetStr
                 // TODO ignore??  schedule async error?
             }
         }
-    }
-
-    // ----------------------------------------------------------------------------------------------------
-    // Request Utils
-    // ----------------------------------------------------------------------------------------------------
-    private Message makeRequest(String subject, byte[] bytes, Duration timeout) throws InterruptedException {
-        return conn.request(prependPrefix(subject), bytes, timeout);
-    }
-
-    private Message makeRequestResponseRequired(String subject, byte[] bytes, Duration timeout) throws IOException {
-        try {
-            return responseRequired(conn.request(prependPrefix(subject), bytes, timeout));
-        } catch (InterruptedException e) {
-            throw new IOException(e);
-        }
-    }
-
-    private Message makeRequestResponseRequired(String subject, Headers headers, byte[] data, boolean utf8mode, Duration timeout, boolean cancelOn503) throws IOException {
-        try {
-            return responseRequired(conn.requestInternal(subject, headers, data, utf8mode, timeout, cancelOn503));
-        } catch (InterruptedException e) {
-            throw new IOException(e);
-        }
-    }
-
-    private Message responseRequired(Message respMessage) throws IOException {
-        if (respMessage == null) {
-            throw new IOException("Timeout or no response waiting for NATS JetStream server");
-        }
-        return respMessage;
-    }
-
-    String prependPrefix(String subject) {
-        return jso.getPrefix() + subject;
-    }
-
-    Duration getRequestTimeout() {
-        return jso.getRequestTimeout();
     }
 }
